# 资料/backend-openapi.yaml
openapi: 3.1.0
info:
  title: Counseling Agent Backend API
  version: 1.2.0-sprint2-mood
  description: |
    心理咨询 Agent 后端接口文档。

    本文档覆盖 Sprint 1 闭环接口，以及 Sprint 2 知识库与情绪趋势 MVP：
    鉴权、文本对话、记忆管理、情绪记录、情绪趋势、安全资源、危机事件、知识搜索、知识详情与知识问答。

    以下脚手架或后续 Sprint 占位能力不纳入本次正式契约：
    tests、voice、完整 RAG。
servers:
  - url: http://localhost:8000
    description: 本地开发环境
...
paths:
...
  /api/v1/chat/threads:
    post:
      tags: [chat]
      summary: 创建会话
      operationId: createThread
      description: 创建 Sprint 1 文本聊天线程。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartThreadRequest"
            example:
              mode: companion
              title: 我想倾诉
      responses:
        "200":
          description: 会话创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartThreadResponse"
              example:
                thread_id: 0f8fad5b-d9cb-469f-a165-70867728950e
                langgraph_thread_id: lg-0f8fad5b-d9cb-469f-a165-70867728950e
                mode: companion
                title: 我想倾诉
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/chat/threads/{thread_id}/messages:
    post:
      tags: [chat]
      summary: 发送消息
      operationId: sendMessage
      description: |
        向指定会话发送一条文本消息，并同步返回 LangGraph 编排后的助手回复。
        当前 Sprint 1 仅要求单次请求-响应模式，不包含流式返回。
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: thread_id
          required: true
          description: 会话 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
            examples:
              empathy:
                summary: 普通倾诉场景
                value:
                  user_id: demo-user
                  content: 我最近压力好大，感觉没人理解我
                  input_type: text
                  user_mode: teen
              crisis:
                summary: 高风险场景
                value:
                  user_id: demo-user
                  content: 我真的不想活了
                  input_type: text
                  user_mode: teen
      responses:
        "200":
          description: 助手回复
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
              examples:
                empathy:
                  summary: 普通共情回复
                  value:
                    thread_id: 0f8fad5b-d9cb-469f-a165-70867728950e
                    assistant_message:
                      assistant_text: I hear you. Thank you for sharing this. We can go slowly and focus on what feels hardest right now.
                      risk_level: L1
                      intent: vent
                      suggested_actions:
                        - continue
                        - summarize
                        - gentle advice
                      session_summary: topic=我最近压力好大，感觉没人理解我 risk=L1
                      should_write_memory: true
                crisis:
                  summary: 高风险安全分流回复
                  value:
                    thread_id: 0f8fad5b-d9cb-469f-a165-70867728950e
                    assistant_message:
                      assistant_text: Your safety matters most right now. I want you to contact a trusted person immediately and use local emergency resources.
                      risk_level: L2
                      intent: other
                      suggested_actions:
                        - contact trusted person
                        - open SOS
                        - local emergency
                      session_summary: topic=我真的不想活了 risk=L2
                      should_write_memory: true
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/chat/threads/{thread_id}/stream:
    post:
      tags: [chat]
      summary: SSE 流式发送消息
      operationId: streamMessage
      description: |
        对话页默认使用该接口接收 text/event-stream。事件顺序通常为
        graph_update -> token* -> final。若首个 SSE 事件前失败，前端可回退到
        `/api/v1/chat/threads/{thread_id}/messages`；若首个事件后中断，前端不得重发同一输入，
        应刷新该 thread 的消息列表以避免重复写入。
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: thread_id
          required: true
          description: 会话 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: SSE 事件流
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events carrying graph_update, token, and final events.
              examples:
                stream:
                  summary: token 与最终动作
                  value: |
                    event: graph_update
                    data: {"node":"risk_classifier","risk_level":"L1"}

                    event: token
                    data: {"text":"我听到了"}

                    event: final
                    data: {"thread_id":"0f8fad5b-d9cb-469f-a165-70867728950e","message_id":"client-msg-001","assistant_message_id":"assistant-msg-001","assistant_text":"我听到了。","risk_level":"L1","intent":"vent","suggested_actions":["继续说","帮我梳理"],"session_summary":"risk=L1","should_write_memory":true}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/memories:
    get:
      tags: [memory]
      summary: 获取记忆列表
      operationId: listMemories
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 记忆列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMemoriesResponse"
              example:
                items:
                  - memory_id: demo-memory-1
                    memory_type: preference
                    content: Prefers gentle responses.
    delete:
      tags: [memory]
      summary: 清空全部记忆
      operationId: clearMemories
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 清空结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
              example:
                status: cleared
  /api/v1/memories/{memory_id}:
    patch:
      tags: [memory]
      summary: 修改单条记忆
      operationId: updateMemory
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: memory_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemoryRequest"
            example:
              content: 用户更偏好简短、温柔的回复
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryMutationResponse"
              example:
                memory_id: demo-memory-1
                content: 用户更偏好简短、温柔的回复
                status: updated
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [memory]
      summary: 删除单条记忆
      operationId: deleteMemory
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: memory_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryMutationResponse"
              example:
                memory_id: demo-memory-1
                status: deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: 未登录或 access token 已失效
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
...
    UserMode:
      type: string
      enum:
        - teen
        - adult
    ThreadMode:
      type: string
      enum:
        - companion
        - knowledge
        - test
        - crisis
      description: |
        当前 Sprint 1 正式使用 `companion` 为主；
        其余值保留字段兼容性，不代表对应能力已纳入 Sprint 1 范围。
    InputType:
      type: string
      enum:
        - text
        - voice
        - test
        - system
    RiskLevel:
      type: string
      enum:
        - L0
        - L1
        - L2
        - L3
      description: |
        L0=正常，L1=中低风险情绪波动，L2/L3=需进入安全分流。
...
    StartThreadRequest:
      type: object
      properties:
        mode:
          $ref: "#/components/schemas/ThreadMode"
          default: companion
        title:
          type: string
      additionalProperties: false
    StartThreadResponse:
      type: object
      properties:
        thread_id:
          type: string
        langgraph_thread_id:
          type: string
        mode:
          type: string
        title:
          type: string
      required:
        - thread_id
        - langgraph_thread_id
        - mode
        - title
    SendMessageRequest:
      type: object
      properties:
        user_id:
          type: string
        content:
          type: string
        input_type:
          $ref: "#/components/schemas/InputType"
          default: text
        user_mode:
          $ref: "#/components/schemas/UserMode"
          default: adult
      required:
        - user_id
        - content
    AssistantMessage:
      type: object
      properties:
        assistant_text:
          type: string
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        intent:
          type: string
          description: LangGraph 意图识别结果
        suggested_actions:
          type: array
          items:
            type: string
        session_summary:
          type: string
          description: 当前轮对话摘要，供首页续聊展示使用
        should_write_memory:
          type: boolean
        referenced_memories:
          type: array
          items:
            $ref: "#/components/schemas/MemoryReference"
      required:
        - assistant_text
        - risk_level
        - intent
        - suggested_actions
        - session_summary
        - should_write_memory
        - referenced_memories
    SendMessageResponse:
      type: object
      properties:
        thread_id:
          type: string
        assistant_message:
          $ref: "#/components/schemas/AssistantMessage"
      required:
        - thread_id
        - assistant_message
...
    MemoryReference:
      type: object
      properties:
        memory_id:
          type: string
        memory_type:
          type: string
        content:
          type: string
      required:
        - memory_id
        - memory_type
        - content
    MemoryItem:
      type: object
      properties:
        memory_id:
          type: string
        memory_type:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - memory_id
        - memory_type
        - content
        - created_at
        - updated_at
    ListMemoriesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MemoryItem"
      required:
        - items
    UpdateMemoryRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    MemoryMutationResponse:
      type: object
      properties:
        memory_id:
          type: string
        content:
          type: string
        status:
          type: string
      required:
        - status
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    ValidationError:
      type: object
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
      required:
        - loc
        - msg
        - type
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
      required:
        - detail